Refactor: Split UVIslandMaskEditor into partial classes
Extract UI Creation code (850 lines) into separate partial class file
to improve maintainability and reduce main file size.

Changes:
- Created UVIslandMaskEditor.UI.cs with all UI creation methods
- Converted UVIslandMaskEditor to partial class
- Main file reduced from 2991 to 2141 lines (28% reduction)

This is the first step toward simplifying the complex lifecycle
management and cache system.
